+++ R/Spectra.R
@@ -381,6 +381,9 @@ NULL
 #'   provided polarity. Returns the filtered `Spectra` (with spectra in their
 #'   original order).
 #'
+#' - `filterPrecursorCharge`: retains spectra with the defined precursor
+#'   charge(s).
+#'
 #' - `filterPrecursorMzRange` (previously `filterPrecursorMz` which is now
 #'   deprecated): retains spectra with a precursor m/z within the
 #'   provided m/z range. See examples for details on selecting spectra with
@@ -390,8 +393,16 @@ NULL
 #'   of the provided m/z values (given `ppm` and `tolerance`). Spectra with
 #'   missing precursor m/z value (e.g. MS1 spectra) are dropped.
 #'
-#' - `filterPrecursorCharge`: retains spectra with the defined precursor
-#'   charge(s).
+#' - `filterPrecursorPeaks`: removes peaks from each spectrum in `object` with
+#'   an m/z equal or larger than the m/z of the precursor, depending on the
+#'   value of parameter `mz`: for `mz = ==" (the default) peaks with matching
+#'   m/z (considering an absolute and relative acceptable difference depending
+#'   on `tolerance` and `ppm`, respectively) are removed. For `mz = ">="` all
+#'   peaks with an m/z larger or equal to the precursor m/z (minus `tolerance`
+#'   and the `ppm` of the precursor m/z) are removed. Parameter `msLevel.`
+#'   allows to restrict the filter to certain MS levels (by default the filter
+#'   is applied to all MS levels). Note that no peaks are removed if the
+#'   precursor m/z is `NA` (e.g. typically for MS1 spectra).
 #'
 #' - `filterPrecursorScan`: retains parent (e.g. MS1) and children scans (e.g.
 #'   MS2) of acquisition number `acquisitionNum`. Returns the filtered
@@ -2253,39 +2264,6 @@ setMethod("filterPrecursorCharge", "Spectra",
                   "Filter: select spectra with a precursor charge ",
                   paste0(z, collapse = ", "))
               object
-          })
-
-#' @rdname Spectra
-setMethod("filterPrecursorPeaks", "Spectra",
-          function(object, mz = numeric(), tolerance = 0, ppm = 20,
-                   msLevel. = uniqueMsLevels(object), remove = c("==", ">=")) {
-              if (!.check_ms_level(object, msLevel.))
-                  return(object)
-              l <- length(mz)
-              if (length(tolerance) != 1)
-                  stop("'tolerance' should be of length 1")
-              if (length(ppm) != 1)
-                  stop("'ppm' should be of length 1")
-              ## filter peaks which:
-              ## - "!=": keep all peaks except those with an m/z similar to the
-              ##         precursor m/z
-              ## - "<": keep all peaks with an m/z < than the precursor m/z
-              ## Maybe don't implement the others yet.
-              ## LLLLL
-              ## object <- addProcessing(object, .peaks_filter_mz_value,
-              ##                         mz = mz, tolerance = tolerance,
-              ##                         ppm = ppm, msLevel = msLevel.,
-              ##                         keep = keep, spectraVariables = "msLevel")
-              ## if (length(mz) <= 3)
-              ##     what <- paste0(format(mz, digits = 4), collapse = ", ")
-              ## else what <- ""
-              ## if (keep)
-              ##     keep_or_remove <- "select"
-              ## else keep_or_remove <- "remove"
-              ## object@processing <- .logging(
-              ##     object@processing, "Filter: ", keep_or_remove,
-              ##     " peaks matching provided m/z values ", what)
-              ## object
           })
 
 #' @rdname Spectra